fix: add more debug logs
