+++ src/job/metrics.rs
@@ -74,7 +74,9 @@ async fn load_ingest_wal_used_bytes() -> Result<(), anyhow::Error> {
     let mut sizes = HashMap::new();
     for file in files {
         let local_file = file.to_owned();
-        let local_path = Path::new(&file).canonicalize().unwrap();
+        let Ok(local_path) = Path::new(&file).canonicalize() else {
+            continue;
+        };
         let file_path = local_path
             .strip_prefix(&data_dir)
             .unwrap()
